requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
cryptography==2.8
azure-storage-blob==12.12.0
joblib
numpy==1.21.6
pandas==1.3.5
scikit-learn==1.0.2
scikit-surprise==1.1.1
